Leasing: 4kv and 10kv bug fixed

+++ kcell-process-app/src/main/java/kz/kcell/flow/leasing/CreateCandidate.java
@@ -1110,7 +1110,7 @@ public class CreateCandidate implements JavaDelegate {
                         ARTEFACT_RR_POWER_PreparedStatement.setNull(i++, Types.FLOAT);
                     }
                     if (res_10kv != null) {
-                        ARTEFACT_RR_POWER_PreparedStatement.setFloat(i++, Integer.parseFloat(res_10kv)); //RES_10KV
+                        ARTEFACT_RR_POWER_PreparedStatement.setFloat(i++, Float.parseFloat(res_10kv)); //RES_10KV
                     } else {
                         ARTEFACT_RR_POWER_PreparedStatement.setNull(i++, Types.FLOAT);
                     }
